[CIDMDEV-4933] Multi-Factor Mobile Passcode :: Generate Bypass Codes :: Feature Flag

+++ src/test/groovy/com/rackspace/idm/multifactor/service/BasicMultiFactorServiceTest.groovy
@@ -19,6 +19,7 @@ import com.rackspace.idm.domain.entity.MobilePhone
 import com.rackspace.idm.domain.entity.OTPDevice
 import com.rackspace.idm.domain.entity.TenantRole
 import com.rackspace.idm.domain.entity.User
+import com.rackspace.idm.exception.BadRequestException
 import com.rackspace.idm.exception.ErrorCodeIdmException
 import com.rackspace.idm.util.BypassHelper
 import com.rackspace.idm.util.OTPHelper
@@ -404,13 +405,16 @@ class BasicMultiFactorServiceTest extends RootServiceTest {
         true            |  true         |  FactorTypeEnum.OTP   | 1                | true
     }
 
-    def "test bypass code logic"() {
+    @Unroll
+    def "test bypass code logic: local=#local, type=#type"() {
         setup:
         staticConfig.getBypassDefaultNumber() >> BigInteger.ONE
         staticConfig.getBypassMaximumNumber() >> BigInteger.TEN
+        reloadableConfig.getFeatureLocalMultifactorBypassEnabled() >> local
+
         User user = entityFactory.createUser().with {
             it.multifactorEnabled = true
-            it.multiFactorType = FactorTypeEnum.OTP.value()
+            it.multiFactorType = type
             return it
         }
         OTPDevice otpDevice = new OTPDevice().with {
@@ -419,25 +423,45 @@ class BasicMultiFactorServiceTest extends RootServiceTest {
             return it
         }
         BypassDevice bypassDevice = new BypassDevice()
-        List<BypassDevice> bypassDevices = [bypassDevice].asList()
-        List<String> fakeCodes = ["1", "2"].asList()
+        def arrayOfCodes = ["1", "2"]
+        List<String> fakeCodes = arrayOfCodes.asList()
 
         userService.checkAndGetUserById(user.id) >> user
         mockOTPDeviceDao.getOTPDeviceByParentAndId(user, otpDevice.id) >> otpDevice
         mockOTPDeviceDao.countVerifiedOTPDevicesByParent(user) >> 1
 
         when: "try to add bypass codes"
-        def codes = service.getSelfServiceBypassCodes(user, 120, 2)
+        def exception = null
+        def codes = null
+        try {
+            codes = service.getSelfServiceBypassCodes(user, 120, 2)
+        } catch (Exception e) {
+            exception = e
+        }
 
         then:
-        1 * mockBypassHelper.createBypassDevice(2, 120) >> bypassDevice
-        1 * mockBypassDeviceDao.deleteAllBypassDevices(user)
-        1 * mockBypassDeviceDao.addBypassDevice(user, bypassDevice)
-        1 * mockBypassHelper.calculateBypassCodes(bypassDevice) >> fakeCodes
+        if (local) {
+            1 * mockBypassHelper.createBypassDevice(2, 120) >> bypassDevice
+            1 * mockBypassDeviceDao.deleteAllBypassDevices(user)
+            1 * mockBypassDeviceDao.addBypassDevice(user, bypassDevice)
+            1 * mockBypassHelper.calculateBypassCodes(bypassDevice) >> fakeCodes
+        } else if (type == FactorTypeEnum.SMS.value()) {
+            1 * multiFactorUserManagement.getBypassCodes(_, _, _) >> arrayOfCodes
+        }
 
         then:
-        codes.size() == 2
-        codes == fakeCodes
+        if (exception == null) {
+            codes.size() == 2
+            codes == fakeCodes
+        } else {
+            exception instanceof BadRequestException
+        }
+
+        where:
+        local | type
+        true  | FactorTypeEnum.OTP.value()
+        false | FactorTypeEnum.OTP.value()
+        false | FactorTypeEnum.SMS.value()
     }
 
 }
